refactor: move assign/get connector session to interaction path
